+++ v2/src/main/java/dev/streamx/cli/v2/commands/settings/get/GetCommand.java
@@ -1,7 +1,7 @@
-package dev.streamx.cli.v2.commands.config.get;
+package dev.streamx.cli.v2.commands.settings.get;
 
-import dev.streamx.cli.v2.commands.config.ConfigCommand;
-import dev.streamx.cli.v2.commands.config.ConfigFile;
+import dev.streamx.cli.v2.commands.settings.SettingsCommand;
+import dev.streamx.cli.v2.commands.settings.SettingsFile;
 import dev.streamx.cli.v2.errors.ErrorPrinter;
 import io.vavr.control.Either;
 import io.vavr.control.Try;
@@ -19,7 +19,7 @@ public class GetCommand implements Runnable {
   private static final Logger logger = Logger.getLogger(GetCommand.class);
 
   @CommandLine.ParentCommand
-  public ConfigCommand configCommand;
+  public SettingsCommand settingsCommand;
 
   @CommandLine.Parameters(index = "0", description = "Property key")
   private String key;
@@ -27,14 +27,14 @@ public class GetCommand implements Runnable {
   @Override
   public void run() {
     printPropertyIfExists(key).mapLeft(e -> {
-      ErrorPrinter.print(logger, e, configCommand.mainCommand.verbose);
+      ErrorPrinter.print(logger, e, settingsCommand.mainCommand.verbose);
       System.exit(1);
       return null;
     });
   }
 
   private Either<RuntimeException, Void> printPropertyIfExists(String key) {
-    return ConfigFile.getUrl()
+    return SettingsFile.getUrl()
       .flatMap(url -> Try.withResources(url::openStream)
         .of(input -> {
           Properties properties = new Properties();
@@ -42,11 +42,11 @@ public class GetCommand implements Runnable {
           return properties;
         })
         .toEither()
-        .mapLeft(e -> new RuntimeException("Unable to load config file", e))
+        .mapLeft(e -> new RuntimeException("Unable to load settings file", e))
         .flatMap(properties -> {
           var value = properties.getProperty(key);
           if (value == null) {
-            return Either.left(new RuntimeException("No such config property found: " + key));
+            return Either.left(new RuntimeException("No such settings property found: " + key));
           }
 
           logger.info(value);
